+++ packages/libs/common/src/yaml-parsing/index.ts
@@ -1,1 +1,2 @@
 export * from "./literate-yaml";
+export * from "./guard";
