+++ packages/libs/common/src/yaml-parsing/literate-yaml.ts
@@ -3,20 +3,11 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import {
-  DataHandle,
-  DataSink,
-  IndexToPosition,
-  ParseNode,
-  createSandbox,
-  StrictJsonSyntaxCheck,
-} from "@azure-tools/datastore";
+import { DataHandle, DataSink, IndexToPosition, ParseNode, StrictJsonSyntaxCheck, Parse } from "@azure-tools/datastore";
 import { OperationAbortedException } from "../exceptions";
 import { AutorestLogger } from "../logging";
-import { mergeYamls, resolveRValue } from "../merging";
+import { identitySourceMapping, resolveRValue, strictMerge } from "../merging";
 import { parseCodeBlocksFromMarkdown } from "./markdown-parser";
-
-const safeEval = createSandbox();
 
 export class CodeBlock {
   info!: string | null;
@@ -31,21 +22,13 @@ export async function parse(logger: AutorestLogger, literate: DataHandle, sink: 
   return parseInternal(logger, literate, sink);
 }
 
-export async function parseCodeBlocks(
-  logger: AutorestLogger,
-  literate: DataHandle,
-  sink: DataSink,
-): Promise<Array<CodeBlock>> {
-  return parseCodeBlocksInternal(logger, literate, sink);
-}
-
 async function parseInternal(logger: AutorestLogger, hLiterate: DataHandle, sink: DataSink): Promise<DataHandle> {
   // merge the parsed codeblocks
-  const blocks = (await parseCodeBlocksInternal(logger, hLiterate, sink)).map((each) => each.data);
+  const blocks = (await parseCodeBlocks(logger, hLiterate, sink)).map((each) => each.data);
   return mergeYamls(logger, blocks, sink);
 }
 
-async function parseCodeBlocksInternal(
+export async function parseCodeBlocks(
   logger: AutorestLogger,
   hLiterate: DataHandle,
   sink: DataSink,
@@ -109,103 +92,49 @@ async function parseCodeBlocksInternal(
   return hsConfigFileBlocks;
 }
 
-export function evaluateGuard(rawFenceGuard: string, contextObject: any, forceAllVersionsMode = false): boolean {
-  // extend the context object so that we can have some helper functions.
-  contextObject = {
-    ...contextObject,
-    /** finds out if there is an extension being loaded already by a given name */
-    isLoaded: (name: string) => {
-      return (
-        contextObject["used-extension"] &&
-        !!contextObject["used-extension"].find((each: any) => each.startsWith(`["${name}"`))
-      );
-    },
+/**
+ * Merge a set of yaml code blocks.
+ * @param logger
+ * @param yamlInputHandles
+ * @param sink
+ */
+export async function mergeYamls(
+  logger: AutorestLogger,
+  yamlInputHandles: DataHandle[],
+  sink: DataSink,
+): Promise<DataHandle> {
+  let mergedGraph: any = {};
+  const mappings = [];
+  const cancel = false;
+  let failed = false;
 
-    /** allows a check to see if a given extension is being requested already */
-    isRequested: (name: string): boolean => {
-      return contextObject["use-extension"]?.[name];
-    },
+  const newIdentity = yamlInputHandles.flatMap((x) => x.identity);
 
-    /** if they are specifying one or more profiles or api-versions, then they are   */
-    enableAllVersionsMode: () => {
-      return forceAllVersionsMode;
-    },
+  for (const yamlInputHandle of yamlInputHandles) {
+    const rawYaml = await yamlInputHandle.ReadData();
+    const inputGraph: any =
+      Parse(rawYaml, (message, index) => {
+        failed = true;
+        if (logger) {
+          logger.trackError({
+            code: "yaml_parsing",
+            message: message,
+            source: [{ document: yamlInputHandle.key, position: IndexToPosition(yamlInputHandle, index) }],
+          });
+        }
+      }) || {};
 
-    /** prints a debug message from configuration. sssshhh. don't use this.  */
-    debugMessage: (text: string) => {
-      // eslint-disable-next-line no-console
-      console.log(text);
-      return true;
-    },
-  };
-
-  // trim the language from the front first
-  let match = /^\S*\s*(.*)/.exec(rawFenceGuard);
-  const fence = match && match[1];
-  if (!fence) {
-    // no fence at all.
-    return true;
+    mergedGraph = strictMerge(mergedGraph, inputGraph);
+    mappings.push(...identitySourceMapping(yamlInputHandle.key, await yamlInputHandle.ReadYamlAst()));
   }
 
-  let guardResult = false;
-  let expressionFence = "";
-  try {
-    if (!fence.includes("$(")) {
-      try {
-        return safeEval<boolean>(fence, contextObject);
-      } catch (e) {
-        //console.log(`1 failed to eval ${fence}`);
-        return false;
-      }
-    }
-
-    expressionFence = `${resolveRValue(fence, "", contextObject, null, 2)}`;
-    // is there unresolved values?  May be old-style. Or the values aren't defined.
-
-    // Let's run it only if there are no unresolved values for now.
-    if (!expressionFence.includes("$(")) {
-      return safeEval<boolean>(expressionFence, contextObject);
-    }
-  } catch (E) {
-    // console.log(`2 failed to eval ${expressionFence}`);
-    // not a legal expression?
+  if (failed) {
+    throw new Error("Syntax errors encountered.");
   }
 
-  // is this a single $( ... ) expression ?
-  match = /^\$\((.*)\)$/.exec(fence.trim());
-
-  const guardExpression = match && !match[1].includes("$(") && match[1];
-  if (!guardExpression) {
-    // Nope. this isn't an old style expression.
-    // at best, it can be an expression that doesn't have all the values resolved.
-    // let's resolve them to undefined and see what happens.
-
-    try {
-      return safeEval<boolean>(expressionFence.replace(/\$\(.*?\)/g, "undefined"), contextObject);
-    } catch {
-      // console.log(`3 failed to eval ${expressionFence.replace(/\$\(.*?\)/g, 'undefined')}`);
-      try {
-        return safeEval<boolean>(fence.replace(/\$\(.*?\)/g, "undefined"), contextObject);
-      } catch {
-        //console.log(`4 failed to eval ${fence.replace(/\$\(.*?\)/g, 'undefined')}`);
-        return false;
-      }
-    }
+  if (cancel) {
+    throw new OperationAbortedException();
   }
 
-  // fall back to original behavior, where the whole expression is in the $( ... )
-  const context = { $: contextObject, ...contextObject };
-
-  try {
-    guardResult = safeEval<boolean>(guardExpression, context);
-  } catch (e) {
-    try {
-      guardResult = safeEval<boolean>("$['" + guardExpression + "']", context);
-    } catch (e) {
-      // at this point, it can only be an single-value expression that isn't resolved
-      // which means return 'false'
-    }
-  }
-
-  return guardResult;
+  return sink.WriteObject("merged YAMLs", mergedGraph, newIdentity, undefined, mappings, yamlInputHandles);
 }
